add apply/types.go
