+++ pkg/apply/apply.go
@@ -25,97 +25,6 @@ import (
 
 var ErrFewerPartitions = errors.New("fewer partitions in topic config")
 
-// TopicApplierConfig contains the configuration for a TopicApplier struct.
-type TopicApplierConfig struct {
-	BrokerThrottleMBsOverride  int
-	BrokersToRemove            []int
-	ClusterConfig              config.ClusterConfig
-	DryRun                     bool
-	JsonOutput                 bool
-	PartitionBatchSizeOverride int
-	Rebalance                  bool
-	AutoContinueRebalance      bool
-	RetentionDropStepDuration  time.Duration
-	SkipConfirm                bool
-	IgnoreFewerPartitionsError bool
-	SleepLoopDuration          time.Duration
-	TopicConfig                config.TopicConfig
-}
-
-// TopicApplier executes an "apply" run on a topic by comparing the actual
-// and desired configurations, and then updating the topic as necessary to
-// align the two.
-type TopicApplier struct {
-	config      TopicApplierConfig
-	adminClient admin.Client
-	brokers     []admin.BrokerInfo
-
-	// Pull out some fields for easier access
-	clusterConfig config.ClusterConfig
-	maxBatchSize  int
-	throttleBytes int64
-	topicConfig   config.TopicConfig
-	topicName     string
-}
-
-// IntValueChanges stores changes in integer values (NumPartitions & ReplicationFactor)
-// if a topic is being created then Updated == Current
-type IntValueChanges struct {
-	Current int
-	Updated int
-}
-
-// ConfigEntryChanges holds configs to be updated, as well as their current and updated values
-// if a topic is being created then Updated == Current
-type ConfigEntryChanges struct {
-	Name    string
-	Current string
-	Updated string
-}
-
-// ReplicaAssignmentChanges stores replica reassignment
-// if a topic is being created then UpdatedReplicas == CurrentReplicas
-// TODO: update Changes to actually support replica reassignment
-type ReplicaAssignmentChanges struct {
-	Partition       int
-	CurrentReplicas []int
-	UpdatedReplicas []int
-}
-
-// enum for possible Action values in ChangesTracker
-type ActionEnum string
-
-const (
-	ActionEnumCreate ActionEnum = "create"
-	ActionEnumUpdate ActionEnum = "update"
-)
-
-// ChangesTracker stores what's changed in a topic during an apply run
-// to eventually be printed to stdout as a JSON blob in subcmd/apply.go
-type ChangesTracker struct {
-	// Topic name
-	Topic string
-
-	// NumPartitions created. -1 indicates unset.
-	NumPartitions IntValueChanges
-
-	// ReplicationFactor for the topic. -1 indicates unset.
-	ReplicationFactor IntValueChanges
-
-	// ReplicaAssignments among kafka brokers for this topic partitions. If this
-	// is set num_partitions and replication_factor must be unset.
-	ReplicaAssignments []ReplicaAssignmentChanges
-
-	// ConfigEntries holds topic level configuration for topic to be set.
-	ConfigEntries []ConfigEntryChanges
-
-	// MissingKeys stores configs which are set in the cluster but not in the topicctl config
-	MissingKeys []string
-
-	// Action records whether this is a topic being created or updated
-	Action ActionEnum
-}
-
 // NewTopicApplier creates and returns a new TopicApplier instance.
 func NewTopicApplier(
 	ctx context.Context,
